requirements.txt
# Core scientific computing
numpy>=1.24
pandas>=2.0
scipy>=1.11

# Visualization
matplotlib>=3.7

# Survival analysis
lifelines>=0.27
scikit-survival>=0.22

# Machine learning
scikit-learn>=1.3
torch>=2.0

# Bioinformatics / GSEA
gseapy>=1.0
statsmodels>=0.14

# Medical imaging
openslide-python>=4.0
pyradiomics>=3.1

# Utilities
requests>=2.31
tqdm>=4.66
